added Class info in info.py to operate infos
