+++ code/init.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from openpyxl import load_workbook
+from info import Info
+
+Info()
